// zilstorage (2)/types.ts
export interface Game {
  id: number;
  name: string;
  image: string;
  description: string;
  launchDate: string;
  developer: string;
  classification: string;
  skills: string;
  obtainLink: string;
}

export type UserStatus = 'Ativo' | 'Bloqueado' | 'Pendente';

export interface User {
  id: number;
  name: string;
  email: string;
  password;
  status: UserStatus;
  isAdmin: boolean;
  favorites: number[];
}

export type Page = 'login' | 'signup' | 'home' | 'about' | 'admin' | 'profile' | 'favorites' | 'help';